chore: move to src/index
